refactor(vt): flatten re-export shim modules
frame.rs, modes.rs, selection.rs, and snapshot.rs were pure pub-use
forwards of seance-frame / seance-protocol items, each adding a module
hop with no logic. lib.rs now re-exports the same crate-root surface
directly and internal users import via the root (RowMeta and ImageId,
which were never part of the root surface, come straight from
seance_protocol). No external user referenced the shim paths.
