--- crates/seance-vt/src/frame.rs
@@ -1,7 +1,0 @@
-pub use seance_frame::{
-    CellView, CellVisitor, FrameSource, ImageInfo, ImageVisitor, PlacementLayer, PlacementVisitor,
-};
-pub use seance_protocol::frame::{
-    CellAttrs, CellColor, CursorInfo, CursorShape, DirtySnapshot, PlacementSnapshot,
-};
-pub use seance_protocol::identity::ImageId;
