--- crates/seance-vt/src/modes.rs
@@ -1,1 +1,0 @@
-pub use seance_protocol::frame::TerminalModes;
