--- crates/seance-vt/src/selection.rs
@@ -1,1 +1,0 @@
-pub use seance_protocol::frame::{GridPos, Selection, SelectionGranularity};
