--- crates/seance-vt/src/snapshot.rs
@@ -1,2 +1,0 @@
-pub use seance_frame::SnapshotFrameSource;
-pub use seance_protocol::frame::{RowMeta, SnapshotCell, SnapshotImage, VtSnapshot};
